refactor: break down vertex.ts into modular components
- Split vertex.ts into focused modules:
  - dvm/cache.ts: DVM cache management
  - dvm/query.ts: DVM query functionality
  - profile/resolution.ts: Author resolution and profile lookup
  - profile/nip05.ts: NIP-05 resolution
  - profile/fallback.ts: Fallback profile lookup
  - profile/verification.ts: NIP-05 verification
  - profile/search.ts: Profile search and ranking
- Maintain backward compatibility through re-exports
- Keep files under 210 lines for better maintainability
